Removed all cvt spaces for reward & application
