+++ pages/[domain]/rewards/applications/[applicationId].tsx
@@ -2,12 +2,9 @@ import { useRouter } from 'next/router';
 
 import getPageLayout from 'components/common/PageLayout/getLayout';
 import { RewardApplicationPage } from 'components/rewards/components/RewardApplicationPage/RewardApplicationPage';
-import { RewardApplicationPageV2 } from 'components/rewards/components/RewardApplicationPage/RewardApplicationPageV2';
-import { useIsCharmverseSpace } from 'hooks/useIsCharmverseSpace';
 
 export default function RewardApplicationPageComponent() {
   const router = useRouter();
-  const isCharmverseSpace = useIsCharmverseSpace();
 
   const applicationId = router.query.applicationId as string;
   const rewardId = router.query.rewardId as string;
@@ -16,13 +13,7 @@ export default function RewardApplicationPageComponent() {
     return null;
   }
 
-  if (isCharmverseSpace) {
-    return <RewardApplicationPageV2 applicationId={applicationId} rewardId={rewardId} />;
-  }
-
-  // TODO - Add event type
-  // useTrackPageView({ type: 'b' });
-  return <RewardApplicationPage applicationId={applicationId} />;
+  return <RewardApplicationPage applicationId={applicationId} rewardId={rewardId} />;
 }
 
 RewardApplicationPageComponent.getLayout = getPageLayout;
